Consolidation of ITOps projects: Communication, Sql, NoSql.
Update of Blue and Black services references to ITOps.
