+++ services/black/src/ambassador/BlackClientFactory.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Net.Http;
 using System.Threading.Tasks;
-using artiso.AdsdHotel.ITOps.Abstraction.NServiceBus;
+using artiso.AdsdHotel.ITOps.Communication.Abstraction.NServiceBus;
 using NServiceBus;
 
 namespace artiso.AdsdHotel.Black.Ambassador
