+++ services/black/src/ambassador/artiso.AdsdHotel.Black.Ambassador.csproj
@@ -14,8 +14,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\..\blue\src\itops\artiso.AdsdHotel.ITOps.csproj" />
-    <ProjectReference Include="..\..\..\infrastructure\src\nservicebus\artiso.AdsdHotel.Infrastructure.NServiceBus.csproj" />
+    <ProjectReference Include="..\..\..\itops\src\communication\artiso.AdsdHotel.ITOps.Communication.csproj" />
     <ProjectReference Include="..\commands\artiso.AdsdHotel.Black.Commands.csproj" />
   </ItemGroup>
