+++ services/black/src/api/Handlers/SetGuestInformationHandler.cs
@@ -4,7 +4,7 @@ using artiso.AdsdHotel.Black.Api.DatabaseModel;
 using artiso.AdsdHotel.Black.Commands;
 using artiso.AdsdHotel.Black.Contracts.Validation;
 using artiso.AdsdHotel.Black.Events;
-using artiso.AdsdHotel.Infrastructure.DataStorage;
+using artiso.AdsdHotel.ITOps.NoSql;
 using DnsClient.Internal;
 using Microsoft.Extensions.Logging;
 using NServiceBus;
