+++ services/blue/artiso.AdsdHotel.Blue.sln
@@ -26,9 +26,11 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "02.api", "02.api", "{4B35E4
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "01.dto", "01.dto", "{AF852D44-26C8-4A80-B504-04B4E0EE0E5F}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "00.itops", "00.itops", "{D5F3E30F-A58C-4CDE-BA54-B7B6FA5D6BAC}"
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "00.itops", "00.itops", "{FD11978A-3F4E-4C57-A9BC-1A04816B018D}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "artiso.AdsdHotel.ITOps", "src\itops\artiso.AdsdHotel.ITOps.csproj", "{1673F218-D266-4A6C-8DCD-AC011FEAB009}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "artiso.AdsdHotel.ITOps.Communication", "..\itops\src\communication\artiso.AdsdHotel.ITOps.Communication.csproj", "{DF7AB353-266F-4E05-8FA2-0C9DEBA4712D}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "artiso.AdsdHotel.ITOps.Sql", "..\itops\src\sql\artiso.AdsdHotel.ITOps.Sql.csproj", "{A29586F2-C132-4C85-9403-F5766E7D4261}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -112,18 +114,30 @@ Global
 		{80386836-10AB-4B43-90B4-A0A337DC577A}.Release|x64.Build.0 = Release|Any CPU
 		{80386836-10AB-4B43-90B4-A0A337DC577A}.Release|x86.ActiveCfg = Release|Any CPU
 		{80386836-10AB-4B43-90B4-A0A337DC577A}.Release|x86.Build.0 = Release|Any CPU
-		{1673F218-D266-4A6C-8DCD-AC011FEAB009}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{1673F218-D266-4A6C-8DCD-AC011FEAB009}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{1673F218-D266-4A6C-8DCD-AC011FEAB009}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{1673F218-D266-4A6C-8DCD-AC011FEAB009}.Debug|x64.Build.0 = Debug|Any CPU
-		{1673F218-D266-4A6C-8DCD-AC011FEAB009}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{1673F218-D266-4A6C-8DCD-AC011FEAB009}.Debug|x86.Build.0 = Debug|Any CPU
-		{1673F218-D266-4A6C-8DCD-AC011FEAB009}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{1673F218-D266-4A6C-8DCD-AC011FEAB009}.Release|Any CPU.Build.0 = Release|Any CPU
-		{1673F218-D266-4A6C-8DCD-AC011FEAB009}.Release|x64.ActiveCfg = Release|Any CPU
-		{1673F218-D266-4A6C-8DCD-AC011FEAB009}.Release|x64.Build.0 = Release|Any CPU
-		{1673F218-D266-4A6C-8DCD-AC011FEAB009}.Release|x86.ActiveCfg = Release|Any CPU
-		{1673F218-D266-4A6C-8DCD-AC011FEAB009}.Release|x86.Build.0 = Release|Any CPU
+		{DF7AB353-266F-4E05-8FA2-0C9DEBA4712D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{DF7AB353-266F-4E05-8FA2-0C9DEBA4712D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{DF7AB353-266F-4E05-8FA2-0C9DEBA4712D}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{DF7AB353-266F-4E05-8FA2-0C9DEBA4712D}.Debug|x64.Build.0 = Debug|Any CPU
+		{DF7AB353-266F-4E05-8FA2-0C9DEBA4712D}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{DF7AB353-266F-4E05-8FA2-0C9DEBA4712D}.Debug|x86.Build.0 = Debug|Any CPU
+		{DF7AB353-266F-4E05-8FA2-0C9DEBA4712D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{DF7AB353-266F-4E05-8FA2-0C9DEBA4712D}.Release|Any CPU.Build.0 = Release|Any CPU
+		{DF7AB353-266F-4E05-8FA2-0C9DEBA4712D}.Release|x64.ActiveCfg = Release|Any CPU
+		{DF7AB353-266F-4E05-8FA2-0C9DEBA4712D}.Release|x64.Build.0 = Release|Any CPU
+		{DF7AB353-266F-4E05-8FA2-0C9DEBA4712D}.Release|x86.ActiveCfg = Release|Any CPU
+		{DF7AB353-266F-4E05-8FA2-0C9DEBA4712D}.Release|x86.Build.0 = Release|Any CPU
+		{A29586F2-C132-4C85-9403-F5766E7D4261}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A29586F2-C132-4C85-9403-F5766E7D4261}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A29586F2-C132-4C85-9403-F5766E7D4261}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{A29586F2-C132-4C85-9403-F5766E7D4261}.Debug|x64.Build.0 = Debug|Any CPU
+		{A29586F2-C132-4C85-9403-F5766E7D4261}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{A29586F2-C132-4C85-9403-F5766E7D4261}.Debug|x86.Build.0 = Debug|Any CPU
+		{A29586F2-C132-4C85-9403-F5766E7D4261}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A29586F2-C132-4C85-9403-F5766E7D4261}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A29586F2-C132-4C85-9403-F5766E7D4261}.Release|x64.ActiveCfg = Release|Any CPU
+		{A29586F2-C132-4C85-9403-F5766E7D4261}.Release|x64.Build.0 = Release|Any CPU
+		{A29586F2-C132-4C85-9403-F5766E7D4261}.Release|x86.ActiveCfg = Release|Any CPU
+		{A29586F2-C132-4C85-9403-F5766E7D4261}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -135,7 +149,8 @@ Global
 		{81DBD6E3-60A7-4E71-8D3C-B2656D098234} = {AF852D44-26C8-4A80-B504-04B4E0EE0E5F}
 		{8931F5D0-49D8-411D-BD19-FEB16A820C32} = {AF852D44-26C8-4A80-B504-04B4E0EE0E5F}
 		{80386836-10AB-4B43-90B4-A0A337DC577A} = {F9773735-4C26-48F0-B419-BBD7505CE9B3}
-		{1673F218-D266-4A6C-8DCD-AC011FEAB009} = {D5F3E30F-A58C-4CDE-BA54-B7B6FA5D6BAC}
+		{DF7AB353-266F-4E05-8FA2-0C9DEBA4712D} = {FD11978A-3F4E-4C57-A9BC-1A04816B018D}
+		{A29586F2-C132-4C85-9403-F5766E7D4261} = {FD11978A-3F4E-4C57-A9BC-1A04816B018D}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {D60899CD-AFD6-45E1-AEE6-D76A99EBFF72}
