+++ services/blue/src/ambassador/BlueAmbassador.cs
@@ -5,7 +5,7 @@ using System.Threading.Tasks;
 using artiso.AdsdHotel.Blue.Commands;
 using artiso.AdsdHotel.Blue.Contracts;
 using artiso.AdsdHotel.Blue.Validation;
-using artiso.AdsdHotel.ITOps.Abstraction;
+using artiso.AdsdHotel.ITOps.Communication.Abstraction;
 using artiso.AdsdHotel.ITOps.Communication;
 
 namespace artiso.AdsdHotel.Blue.Ambassador
