+++ services/blue/src/api/Handlers/RoomTypeSelectionHandler.cs
@@ -6,7 +6,7 @@ using artiso.AdsdHotel.Blue.Contracts;
 using artiso.AdsdHotel.Blue.Events;
 using artiso.AdsdHotel.Blue.Validation;
 using artiso.AdsdHotel.ITOps.Communication;
-using artiso.AdsdHotel.ITOps.Database.Sql;
+using artiso.AdsdHotel.ITOps.Sql;
 using Microsoft.Extensions.Logging;
 using NServiceBus;
 using RepoDb;
