+++ services/blue/src/api/HostBuilderConfigurationExtensions.cs
@@ -1,4 +1,4 @@
-﻿using artiso.AdsdHotel.ITOps.Abstraction.NServiceBus;
+﻿using artiso.AdsdHotel.ITOps.Communication.Abstraction.NServiceBus;
 using Microsoft.Extensions.Hosting;
 using NServiceBus;
