--- services/infrastructure/artiso.AdsdHotel.Infrastructure.sln
@@ -1,65 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.31005.135
-MinimumVisualStudioVersion = 15.0.26124.0
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "artiso.AdsdHotel.Infrastructure.DataStorage", "src\dataStorage\artiso.AdsdHotel.Infrastructure.DataStorage.csproj", "{97BBBC70-6298-4CC3-A920-EF1555953586}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "artiso.AdsdHotel.Infrastructure.MongoDataStorage", "src\mongoDataStorage\artiso.AdsdHotel.Infrastructure.MongoDataStorage.csproj", "{C1265718-7B89-4CD4-AAF6-B671558F2D97}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "artiso.AdsdHotel.Infrastructure.NServiceBus", "src\nservicebus\artiso.AdsdHotel.Infrastructure.NServiceBus.csproj", "{175E7450-2CB1-4A86-8CDB-73E8D45C5C5E}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|Any CPU = Release|Any CPU
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{97BBBC70-6298-4CC3-A920-EF1555953586}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{97BBBC70-6298-4CC3-A920-EF1555953586}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{97BBBC70-6298-4CC3-A920-EF1555953586}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{97BBBC70-6298-4CC3-A920-EF1555953586}.Debug|x64.Build.0 = Debug|Any CPU
-		{97BBBC70-6298-4CC3-A920-EF1555953586}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{97BBBC70-6298-4CC3-A920-EF1555953586}.Debug|x86.Build.0 = Debug|Any CPU
-		{97BBBC70-6298-4CC3-A920-EF1555953586}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{97BBBC70-6298-4CC3-A920-EF1555953586}.Release|Any CPU.Build.0 = Release|Any CPU
-		{97BBBC70-6298-4CC3-A920-EF1555953586}.Release|x64.ActiveCfg = Release|Any CPU
-		{97BBBC70-6298-4CC3-A920-EF1555953586}.Release|x64.Build.0 = Release|Any CPU
-		{97BBBC70-6298-4CC3-A920-EF1555953586}.Release|x86.ActiveCfg = Release|Any CPU
-		{97BBBC70-6298-4CC3-A920-EF1555953586}.Release|x86.Build.0 = Release|Any CPU
-		{C1265718-7B89-4CD4-AAF6-B671558F2D97}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{C1265718-7B89-4CD4-AAF6-B671558F2D97}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{C1265718-7B89-4CD4-AAF6-B671558F2D97}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{C1265718-7B89-4CD4-AAF6-B671558F2D97}.Debug|x64.Build.0 = Debug|Any CPU
-		{C1265718-7B89-4CD4-AAF6-B671558F2D97}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{C1265718-7B89-4CD4-AAF6-B671558F2D97}.Debug|x86.Build.0 = Debug|Any CPU
-		{C1265718-7B89-4CD4-AAF6-B671558F2D97}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{C1265718-7B89-4CD4-AAF6-B671558F2D97}.Release|Any CPU.Build.0 = Release|Any CPU
-		{C1265718-7B89-4CD4-AAF6-B671558F2D97}.Release|x64.ActiveCfg = Release|Any CPU
-		{C1265718-7B89-4CD4-AAF6-B671558F2D97}.Release|x64.Build.0 = Release|Any CPU
-		{C1265718-7B89-4CD4-AAF6-B671558F2D97}.Release|x86.ActiveCfg = Release|Any CPU
-		{C1265718-7B89-4CD4-AAF6-B671558F2D97}.Release|x86.Build.0 = Release|Any CPU
-		{175E7450-2CB1-4A86-8CDB-73E8D45C5C5E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{175E7450-2CB1-4A86-8CDB-73E8D45C5C5E}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{175E7450-2CB1-4A86-8CDB-73E8D45C5C5E}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{175E7450-2CB1-4A86-8CDB-73E8D45C5C5E}.Debug|x64.Build.0 = Debug|Any CPU
-		{175E7450-2CB1-4A86-8CDB-73E8D45C5C5E}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{175E7450-2CB1-4A86-8CDB-73E8D45C5C5E}.Debug|x86.Build.0 = Debug|Any CPU
-		{175E7450-2CB1-4A86-8CDB-73E8D45C5C5E}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{175E7450-2CB1-4A86-8CDB-73E8D45C5C5E}.Release|Any CPU.Build.0 = Release|Any CPU
-		{175E7450-2CB1-4A86-8CDB-73E8D45C5C5E}.Release|x64.ActiveCfg = Release|Any CPU
-		{175E7450-2CB1-4A86-8CDB-73E8D45C5C5E}.Release|x64.Build.0 = Release|Any CPU
-		{175E7450-2CB1-4A86-8CDB-73E8D45C5C5E}.Release|x86.ActiveCfg = Release|Any CPU
-		{175E7450-2CB1-4A86-8CDB-73E8D45C5C5E}.Release|x86.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {9E53BF01-4886-4A75-97F9-E86E8345E911}
-	EndGlobalSection
-EndGlobal
