--- services/infrastructure/src/dataStorage/artiso.AdsdHotel.Infrastructure.DataStorage.csproj
@@ -1,3 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-</Project>
