--- services/infrastructure/src/mongoDataStorage/artiso.AdsdHotel.Infrastructure.MongoDataStorage.csproj
@@ -1,11 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <ItemGroup>
-    <PackageReference Include="MongoDB.Driver" Version="2.11.6" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\dataStorage\artiso.AdsdHotel.Infrastructure.DataStorage.csproj" />
-  </ItemGroup>
-
-</Project>
