--- services/infrastructure/src/nservicebus/artiso.AdsdHotel.Infrastructure.NServiceBus.csproj
@@ -1,10 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <ItemGroup>
-    <PackageReference Include="NServiceBus" Version="$(Nuget-NServiceBus-Version)" />
-    <PackageReference Include="NServiceBus.Newtonsoft.Json" Version="$(Nuget-NServiceBus-NewtonSoft-Version)" />
-    <PackageReference Include="NServiceBus.RabbitMQ" Version="$(Nuget-NServiceBus-RabbitMQ-Version)" />
-    <PackageReference Include="NServiceBus.Callbacks" Version="$(Nuget-NServiceBus-Callbacks-Version)" />
-  </ItemGroup>
-
-</Project>
