+++ services/itops/artiso.AdsdHotel.ITOps.sln
@@ -1,0 +1,37 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.31019.35
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "artiso.AdsdHotel.ITOps.Communication", "src\communication\artiso.AdsdHotel.ITOps.Communication.csproj", "{378A2AE5-CF0C-4055-8C84-00C38CA55690}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "artiso.AdsdHotel.ITOps.NoSql", "src\nosql\artiso.AdsdHotel.ITOps.NoSql.csproj", "{6E00FB5A-CE53-4005-8120-9E3EC840CFCC}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "artiso.AdsdHotel.ITOps.Sql", "src\sql\artiso.AdsdHotel.ITOps.Sql.csproj", "{CCCFD80B-A319-46DF-8FED-596D588F10A1}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{378A2AE5-CF0C-4055-8C84-00C38CA55690}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{378A2AE5-CF0C-4055-8C84-00C38CA55690}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{378A2AE5-CF0C-4055-8C84-00C38CA55690}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{378A2AE5-CF0C-4055-8C84-00C38CA55690}.Release|Any CPU.Build.0 = Release|Any CPU
+		{6E00FB5A-CE53-4005-8120-9E3EC840CFCC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6E00FB5A-CE53-4005-8120-9E3EC840CFCC}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6E00FB5A-CE53-4005-8120-9E3EC840CFCC}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6E00FB5A-CE53-4005-8120-9E3EC840CFCC}.Release|Any CPU.Build.0 = Release|Any CPU
+		{CCCFD80B-A319-46DF-8FED-596D588F10A1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{CCCFD80B-A319-46DF-8FED-596D588F10A1}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{CCCFD80B-A319-46DF-8FED-596D588F10A1}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{CCCFD80B-A319-46DF-8FED-596D588F10A1}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {0CA80B6C-3B84-4C1D-BF08-B1D292E76826}
+	EndGlobalSection
+EndGlobal
