+++ services/itops/src/communication/Abstraction/NServiceBus/EndpointHolder.cs
@@ -2,7 +2,7 @@
 using System.Threading.Tasks;
 using NServiceBus;
 
-namespace artiso.AdsdHotel.ITOps.Abstraction.NServiceBus
+namespace artiso.AdsdHotel.ITOps.Communication.Abstraction.NServiceBus
 {
     public class EndpointHolder : IAsyncDisposable
     {
