+++ services/itops/src/communication/Abstraction/NServiceBus/NServiceBusEndpointConfigurationFactory.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using NServiceBus;
 
-namespace artiso.AdsdHotel.ITOps.Abstraction.NServiceBus
+namespace artiso.AdsdHotel.ITOps.Communication.Abstraction.NServiceBus
 {
     public class NServiceBusEndpointConfigurationFactory
     {
