+++ services/itops/src/nosql/artiso.AdsdHotel.ITOps.NoSql.csproj
@@ -1,0 +1,7 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <ItemGroup>
+    <PackageReference Include="MongoDB.Driver" Version="2.11.6" />
+  </ItemGroup>
+    
+</Project>
