+++ services/itops/src/sql/DatabaseExtensions.cs
@@ -3,7 +3,7 @@ using System.Data.Common;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace artiso.AdsdHotel.ITOps.Database.Sql
+namespace artiso.AdsdHotel.ITOps.Sql
 {
     public static class DatabaseExtensions
     {
