+++ services/itops/src/sql/IDbConnectionFactory.cs
@@ -1,7 +1,7 @@
 ﻿using System.Data.Common;
 using System.Threading.Tasks;
 
-namespace artiso.AdsdHotel.ITOps.Database.Sql
+namespace artiso.AdsdHotel.ITOps.Sql
 {
     public interface IDbConnectionFactory
     {
