+++ services/itops/src/sql/MySqlConnectionFactory.cs
@@ -2,7 +2,7 @@
 using System.Threading.Tasks;
 using MySql.Data.MySqlClient;
 
-namespace artiso.AdsdHotel.ITOps.Database.Sql
+namespace artiso.AdsdHotel.ITOps.Sql
 {
     public class MySqlConnectionFactory : IDbConnectionFactory
     {
